fix: remove unused files + fix import
