+++ micromanager_gui/_gui_objects/_mda_widget/_mda_widget.py
@@ -4,11 +4,11 @@ import warnings
 from pathlib import Path
 from typing import TYPE_CHECKING
 
+from pymmcore_widgets._core import get_core_singleton
 from qtpy import QtWidgets as QtW
 from qtpy.QtCore import Qt
 from useq import MDASequence
 
-from ..._core import get_core_singleton
 from ..._mda import SEQUENCE_META, SequenceMeta
 from ._mda_gui import MultiDWidgetGui
