+++ micromanager_gui/_gui_objects/_xyz_stages.py
@@ -1,12 +1,10 @@
 from typing import Optional
 
 from pymmcore_plus import DeviceType
-from pymmcore_widgets._stage_widget import StageWidget
+from pymmcore_widgets import StageWidget, _core
 from qtpy.QtCore import QMimeData, Qt
 from qtpy.QtGui import QDrag
 from qtpy.QtWidgets import QGroupBox, QHBoxLayout, QWidget
-
-from .. import _core
 
 STAGE_DEVICES = {DeviceType.Stage, DeviceType.XYStage}
