+++ micromanager_gui/_util.py
@@ -1,9 +1,11 @@
 from __future__ import annotations
 
+import re
 from pathlib import Path
 from typing import TYPE_CHECKING, Iterator, List, Optional, Sequence, Tuple
 
 from pymmcore_plus import CMMCorePlus
+from pymmcore_widgets._core import get_core_singleton
 from qtpy.QtCore import Signal
 from qtpy.QtWidgets import (
     QComboBox,
@@ -16,10 +18,11 @@ from qtpy.QtWidgets import (
     QWidget,
 )
 
-from ._core import get_core_singleton
-
 if TYPE_CHECKING:
     import useq
+
+MAG_PATTERN = re.compile(r"(\d{1,3})[xX]")
+RESOLUTION_ID_PREFIX = "px_size_"
 
 
 def ensure_unique(path: Path, extension: str = ".tif", ndigits: int = 3) -> Path:
@@ -133,3 +136,48 @@ def get_group_dev_prop(
     for preset in mmc.getAvailableConfigs(group):
         dev_props.extend([(k[0], k[1]) for k in mmc.getConfigData(group, preset)])
     return dev_props
+
+
+def update_pixel_size(
+    pixel_size: float,
+    obj_dev_label: Optional[str] = None,
+    mmc: Optional[CMMCorePlus] = None,
+):
+    """Update the pixel size config for objective device `obj_dev_label`.
+
+    Parameters
+    ----------
+    pixel_size : float
+        The camera physical pixel size in microns.
+    obj_dev_label : Optional[str]
+        Device label of the objective to update.  If not provided, will use the
+        current `objective_device` label from `CoreState`.`
+    mmc : Optional[CMMCorePlus]
+        optional mmcore object, by default `CMMCorePlus.instance()`
+    """
+    if pixel_size == 1.0:
+        # TODO: this was previous behavior, but does this make sense?
+        # presumably it was here because the default value of the spinbox is 1.0?
+        return
+
+    mmc = mmc or get_core_singleton()
+    # if pixel size is already set, and we're not providing a new value, return.
+    if current_px_size_cfg := mmc.getCurrentPixelSizeConfig():
+        if not pixel_size:
+            return
+        mmc.deletePixelSizeConfig(current_px_size_cfg)
+
+    # create and store a new pixel size config for the current objective.
+    if not obj_dev_label:
+        return
+    curr_obj = mmc.getProperty(obj_dev_label, "Label")
+
+    # get magnification info from the current objective label
+    if match := MAG_PATTERN.search(curr_obj):
+        mag = int(match.groups()[0])
+
+        # set image pixel sixe (x,y) for the newly created pixel size config
+        resolutionID = f"{RESOLUTION_ID_PREFIX}{curr_obj}"
+        mmc.definePixelSizeConfig(resolutionID, obj_dev_label, "Label", curr_obj)
+        mmc.setPixelSizeUm(resolutionID, pixel_size / mag)
+        mmc.setPixelSizeConfig(resolutionID)
