+++ setup.cfg
@@ -34,6 +34,7 @@ install_requires =
     superqt >=0.3.1
     fonticon-materialdesignicons6
     tifffile
+    zarr
     pymmcore-widgets @ git+https://github.com/pymmcore-plus/pymmcore-widgets.git
 
 [options.extras_require]
