+++ tests/conftest.py
@@ -4,7 +4,6 @@ from typing import Tuple
 import pytest
 from pymmcore_plus import CMMCorePlus
 
-from micromanager_gui import _core
 from micromanager_gui._gui_objects._sample_explorer_widget._sample_explorer_widget import (  # noqa: E501
     MMExploreSample,
 )
@@ -15,15 +14,18 @@ ExplorerTuple = Tuple[MainWindow, MMExploreSample]
 
 @pytest.fixture(params=["local"])
 def global_mmcore(request):
-    _core._SESSION_CORE = CMMCorePlus()  # refresh singleton
-    if request.param == "remote":
-        from pymmcore_plus import server
 
-        server.try_kill_server()
+    mmc = CMMCorePlus()
+    mmc.loadSystemConfiguration(str(Path(__file__).parent / "test_config.cfg"))
+    # _core._SESSION_CORE = CMMCorePlus()  # refresh singleton
+    # if request.param == "remote":
+    #     from pymmcore_plus import server
 
-    mmc = _core.get_core_singleton(remote=request.param == "remote")
-    if len(mmc.getLoadedDevices()) < 2:
-        mmc.loadSystemConfiguration(str(Path(__file__).parent / "test_config.cfg"))
+    #     server.try_kill_server()
+
+    # mmc = _core.get_core_singleton(remote=request.param == "remote")
+    # if len(mmc.getLoadedDevices()) < 2:
+    #     mmc.loadSystemConfiguration(str(Path(__file__).parent / "test_config.cfg"))
     return mmc
